"Fixed" --help option. Now prints to a visible console.
The "--help" option will now print helpful output to the appropriate console window, and exit the program afterwards. There is a bug described in the code that leads to a cluttered console window. I'm not aware of any solutions to this except removing the "--help" option.

Modified the usage output from the previous version to provide a explanation for the color parameter. Issue #7 seems like the most user-friendly way to select colors, and will be the next step.