Explicit unicode support and command-line arguments. (#6)
* Unicodeifying and whitespace fixing

* Command-line options

Implemented options for changing to transparent or opaque appearance,
and an option for specifying which color should be applied.

* Forgot to change LoadLibrary to LoadLibraryW.

* Renamed field and added more descriptive constants 

Renamed OPTIONS.mode to OPTIONS.taskbar_appearance and added constants with more descriptive names to match this field. Modified SetWindowBlur to use the ACCENT_* constants instead of literals when assigning the the 'policy' variable.

* "Fixed" --help option. Now prints to a visible console.

The "--help" option will now print helpful output to the appropriate console window, and exit the program afterwards. There is a bug described in the code that leads to a cluttered console window. I'm not aware of any solutions to this except removing the "--help" option.

Modified the usage output from the previous version to provide a explanation for the color parameter. Issue #7 seems like the most user-friendly way to select colors, and will be the next step.

* Create usage.md